#691: Portfolio pulse on price change
